+++ source/quaternion.h
@@ -7,31 +7,31 @@ namespace LibRayMarching
 
 	struct Quaternion
 	{
-		double r;
-		double i;
-		double j;
-		double k;
+		float r;
+		float i;
+		float j;
+		float k;
 
 		Quaternion();
 		Quaternion(const Quaternion& Quaternion);
-		Quaternion(const double& r, const double& i, const double& j, const double& k);
+		Quaternion(const float& r, const float& i, const float& j, const float& k);
 
 		Quaternion operator+(const Quaternion& quat) const;
 		Quaternion operator-(const Quaternion& quat) const;
 		Quaternion operator*(const Quaternion& quat) const;
-		Quaternion operator*(const double& val) const;
+		Quaternion operator*(const float& val) const;
 		Quaternion operator/(const Quaternion& quat) const;
-		Quaternion operator/(const double& val) const;
+		Quaternion operator/(const float& val) const;
 
 		Quaternion operator+=(const Quaternion& quat);
 		Quaternion operator-=(const Quaternion& quat);
 		Quaternion operator*=(const Quaternion& quat);
-		Quaternion operator*=(const double& val);
-		Quaternion operator/=(const double& val);
+		Quaternion operator*=(const float& val);
+		Quaternion operator/=(const float& val);
 
-		double length() const;
-		double norm() const;
-		double dot(const Quaternion& quat) const;
+		float length() const;
+		float norm() const;
+		float dot(const Quaternion& quat) const;
 		Quaternion cross(const Quaternion& quat) const;
 		Quaternion squared() const;
 		void normalize();
